Hoonmain6

+++ src/main/java/com/onshop/shop/article/ArticleService.java
@@ -7,14 +7,14 @@ public interface ArticleService {
     List<ArticleDTO> getAllArticles();
 
     // ✅ 특정 게시물 조회
-    ArticleDTO getArticleById(int id);
+    ArticleDTO getArticleById(Long id);
 
     // ✅ 게시물 생성
     ArticleDTO createArticle(ArticleDTO articleDTO);
 
     // ✅ 게시물 수정 (부분 업데이트 지원)
-    ArticleDTO updateArticle(int id, ArticleDTO articleDTO);
+    ArticleDTO updateArticle(Long id, ArticleDTO articleDTO);
 
     // ✅ 게시물 삭제
-    void deleteArticle(int id);
+    void deleteArticle(Long id);
 }
